transcoder_container/requirements.txt
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-queue==12.12.0
bump2version==1.0.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
google-api-core==2.24.2
google-auth==2.40.1
googleapis-common-protos==1.70.0
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
